+++ src/main/java/POJOs/Player.java
@@ -4,23 +4,23 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Player {
-    private final int score;
-    private final int playerLeft;
+    private final long score;
+    private final long playerLeft;
     private final String userName;
     private final PlayerAnswers answers;
 
-    public Player(int score, int playerLeft, String userName, PlayerAnswers answers) {
+    public Player(long score, long playerLeft, String userName, PlayerAnswers answers) {
         this.score = score;
         this.playerLeft = playerLeft;
         this.userName = userName;
         this.answers = answers;
     }
 
-    public int getScore() {
+    public long getScore() {
         return score;
     }
 
-    public int getPlayerLeft() {
+    public long getPlayerLeft() {
         return playerLeft;
     }
